Make HTML table converters recursive
We add a base class HTMLNotebookOutputConverter which extends
NotebookOutputConverter. This type of NotebookOutputConverter
specifically produces some kind of HTMLNotebookOutput.

Then, we update the Map<K,V> and List<Map<K,V>> converters to extend
this base class, so they gain access to its convenience asHTML(Object)
method, which itself leans on the ConvertService to convert each table
cell into HTML.

In this way, we can now display tables of images and so forth.