Add error handling to convert statements
This will prevent the kernel from appearing to hang when no suitable
converter is found. Additionally, this commit returns error message when
a converter is not found or when there are no outputs. Null is not
returned for no output scenarios, in order to differ these from actual
null outputs.