Revert "Update to latest beakerx base kernel"
This reverts commit 7f0dbe8efaa325794d2a7b4a6d6dc7b70e7e21ce.

The updated version of beakerx causes the below error:
[Thread-2] INFO com.twosigma.beaker.jupyter.msg.MessageCreator -
Execution result ERROR: The kernel was not ready, run the cell again.

This error occurs whenever you try to run a cell, even though the
scijava-jupyter kernel says it is ready to use. Running the cell again
does not work. Reverting back to the previous version of beakerx fixes
the problem.